pyproject.toml
[project]
name = "React_agent"
version = "0.1.0"
description = "A ReAct agent built with Google ADK"
requires-python = ">=3.11"

dependencies = [
    "google-adk",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/React_agent"]